Expose PSQL parser helpers in new 'Internal' module
This is part of the followup to #1298.

The code exposed in this new `Database.Persist.Postgresql.Internal`
module is very helpful to third-party libraries which want to use
functions from `postgresql-simple` to produce `PersistValue`s for
feeding to `persistent` functions.

Exposing them is necessary to allow for a third-party library which
provides streaming of Persistent entities using Postgres cursors.
